% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grmd.R
\docType{package}
\name{Grmd-package}
\alias{Grmd-package}
\title{A package for improving publishing through word-processor integration}
\description{
The new rmarkdown-package developed by and included in RStudio
allows for simplifications of the porting to word documents.
Apart from having a direct port it allows tailoring of formatters
that make the conversion between html and MS Word virtually seemless.
}
