% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Grmd}
\alias{Grmd}
\alias{Grmd-package}
\title{A package for improving publishing through word-processor integration}
\description{
The new rmarkdown-package developed by and included in
RStudio allows for simplifications of the porting to word
documents. Apart from having a direct port it allows
tailoring of formatters that make the conversion between
html and MS Word virtually seemless.
}

