% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prCaptionFix}
\alias{prCaptionFix}
\title{Fixes the caption back fro elements withou}
\usage{
prCaptionFix(outFile)
}
\arguments{
  \item{outFile}{The name of the file}
}
\value{
outFile The name of the file
}
\description{
Fixes the caption back fro elements withou
}
\keyword{internal}

