% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prFtpOtherChanges}
\alias{prFtpOtherChanges}
\title{Changes lines for XML-conformity}
\usage{
prFtpOtherChanges(output_str)
}
\arguments{
  \item{output_str}{The input from readLines()}
}
\value{
string Returns with changes
}
\description{
Changes lines for XML-conformity
}
\keyword{internal}

